Lots more functionality
Add support for controlling the enclosure LEDs (--led) and
Exhaust fan (--fan). Both take an integer in the range,
0 (off)...100 (max brightness).

Add support for moving to a specific (x,y,z) location. For now,
after --home, these coordinates are machine default coordinates.

To force the head to move to that location, we've added the --move
command line flag. Also, to view the current head location, we've added
the --locate flag. This executes before any --move.

If the user requests a --photo (photo.jpg), and provides a --marks
argument, the photo will be augmented with a magenta X to help show
the cetern pixels of the photo. The intention is that this can
eventually help target the laser.

The --zoom and --circle commands now operates with respect to the
(x,y,z) and steps closer to the base by --zd increments for each
successive position.

The --spot and --nospot commands can attempt to enable the laser for
the duration of a --photo.

Signed-off-by: Tinkerer <tinkerer@zappem.net>