1.08 is a ratio that works for my 1.6W laser
Likely not a reliable relationship, but this is a step towards
recognizing 10mm separation with the camera. If we can identify
two things that are 10mm apart in the X and Y directions, then
we can use the image of those points with the --mark overlay to
recognize when we are focused on one point and the other implies
we are at the right distance (in the Z direction) from the first
point.

Signed-off-by: Tinkerer <tinkerer@zappem.net>