Support minor rewriting of g-code files
I had a CNC drill bit break mid-program and wanted to restart it
from where it stopped drilling. It took a while to re-discover the
origin, and I wanted to minimize the remaining job time. The --edit
flag can be used to comment out specific (ranges of) lines in the
file. It isn't friendly to use, but was enough to recover my job
with a new drill bit.

Signed-off-by: Tinkerer <tinkerer@zappem.net>