More features
Explored some of the other printer heads. More work needed on that front.
Added support for remembering the offset of the camera to the laser
focal point, and introduced --snap which will move to that offset, take
a picture and then return to the original head position. This is really
convenient for using the camera to adjust the head location, and can
be helpful in preparation for --set-origin.

Signed-off-by: Tinkerer <tinkerer@zappem.net>